Convert the difficulty integer parameter to float.

+++ Assets/Scripts/DifficultyButton.cs
@@ -7,7 +7,7 @@ public class DifficultyButton : MonoBehaviour
 {
     private Button _button;
     private GameManager _gameManager;
-    public int difficulty;
+    public float difficulty;
 
     // Start is called before the first frame update
     void Start()
